Factorize code
